requirements.txt
sounddevice
scipy
numpy
groq
python-dotenv
pyttsx3
